requirements.txt
netaddr==0.8.0
pythonping==1.1.4
urllib3==1.26.14
certifi==2022.12.7
python-netbox==0.0.22
